feat: refactoring components
