+++ src/App.css
@@ -1,28 +1,3 @@
 .App {
   font-size: 18px;
 }
-
-.movieTitle {
-  display: block;
-  border: 1px solid gray;
-  border-radius: 5px;
-  width: 500px;
-  color: black;
-  text-decoration: none;
-}
-
-.movieTitle:hover {
-  background: gray;
-}
-
-.good {
-  color: blue;
-}
-
-.soso {
-  color: green;
-}
-
-.bad {
-  color: red;
-}
