+++ src/components/MovieList/style.css
@@ -1,0 +1,26 @@
+
+.movieTitle {
+    display: block;
+    border: 1px solid gray;
+    border-radius: 5px;
+    width: 500px;
+    color: black;
+    text-decoration: none;
+  }
+  
+  .movieTitle:hover {
+    background: gray;
+  }
+  
+  .good {
+    color: blue;
+  }
+  
+  .soso {
+    color: green;
+  }
+  
+  .bad {
+    color: red;
+  }
+
