+++ src/components/Title/index.js
@@ -1,7 +1,8 @@
 import React from 'react'
+import './style.css'
 
 export default function Title() {
     return (
-        <h1>무비 앱</h1>
+        <h1 className="appTitle">무비 앱</h1>
     )
 }
