+++ src/components/Title/style.css
@@ -1,0 +1,4 @@
+.appTitle {
+    font-size: 35px;
+    text-align: center;
+}
